# specs/003-ai-chatbot-leads/contracts/ai-chatbot-leads.openapi.yaml
openapi: 3.1.0
info:
  title: Digital Marketing Hub - AI Chatbot, Knowledge Base, and Lead Capture
  version: 0.1.0
  description: Public chatbot, lead capture, admin knowledge base, admin lead management, notification, and abuse-limit APIs.
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
...
paths:
...
  /admin/leads:
    get:
      tags: [Admin Leads]
      summary: List leads for admin follow-up
      operationId: listAdminLeads
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/LeadStatus"
        - name: source
          in: query
          schema:
            $ref: "#/components/schemas/LeadSource"
      responses:
        "200":
          description: Admin lead list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadListResponse"
  /admin/leads/{id}:
    get:
      tags: [Admin Leads]
      summary: Read lead detail
      operationId: getAdminLead
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Lead detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadDetailResponse"
    patch:
      tags: [Admin Leads]
      summary: Update lead details
      operationId: updateAdminLead
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLeadUpdateRequest"
      responses:
        "200":
          description: Lead updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadDetailResponse"
    delete:
      tags: [Admin Leads]
      summary: Delete lead from normal workflows
      operationId: deleteAdminLead
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          $ref: "#/components/responses/Success"
  /admin/leads/{id}/status:
    patch:
      tags: [Admin Leads]
      summary: Update lead status
      operationId: updateLeadStatus
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  $ref: "#/components/schemas/LeadStatus"
      responses:
        "200":
          description: Lead status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadDetailResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
...
  responses:
    Success:
      description: Operation completed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SuccessResponse"
...
  schemas:
    SuccessResponse:
      type: object
      required: [success, message, data]
      properties:
        success:
          type: boolean
          const: true
        message:
          type: string
        data:
          type: object
          additionalProperties: true
...
    LeadSource:
      type: string
      enum: [CHATBOT, CONTACT_FORM, SERVICE_PAGE, BLOG_PAGE, MANUAL]
    LeadStatus:
      type: string
      enum: [NEW, CONTACTED, QUALIFIED, PROPOSAL_SENT, WON, LOST]
...
    PublicLeadCreateRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        companyName:
          type: string
        serviceInterest:
          type: string
        budgetRange:
          type: string
        message:
          type: string
        source:
          $ref: "#/components/schemas/LeadSource"
    AdminLeadUpdateRequest:
      allOf:
        - $ref: "#/components/schemas/PublicLeadCreateRequest"
        - type: object
          properties:
            assignedTo:
              type: string
              format: uuid
              nullable: true
...
    Lead:
      type: object
      required: [id, source, status, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        companyName:
          type: string
          nullable: true
        serviceInterest:
          type: string
          nullable: true
        budgetRange:
          type: string
          nullable: true
        message:
          type: string
          nullable: true
        source:
          $ref: "#/components/schemas/LeadSource"
        status:
          $ref: "#/components/schemas/LeadStatus"
        assignedTo:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        notifications:
          type: array
          items:
            $ref: "#/components/schemas/LeadNotification"
    LeadNotification:
      type: object
      required: [id, leadId, target, status, payload, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        leadId:
          type: string
          format: uuid
        target:
          type: string
        status:
          type: string
          enum: [PENDING, SENT, FAILED, SKIPPED]
        payload:
          type: object
          additionalProperties: true
        errorMessage:
          type: string
          nullable: true
        attemptedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    LeadListResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessResponse"
        - type: object
          properties:
            data:
              type: object
              required: [items, meta]
              properties:
                items:
                  type: array
                  items:
                    $ref: "#/components/schemas/Lead"
                meta:
                  type: object
                  additionalProperties: true
    LeadDetailResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/Lead"